Update ARM and x86 kernel 3.4 images.
This updates the prebuilt binaries of the 3.4 kernels for
ARM and x86 to commit d3ffcc7b914fb6f91e6e4fb9d6453e778f8cd508

This is required to get rid of a kernel panic, related to cgroups,
that occurs while running CTS tests on the emulator.

NOTE: Mips kernel is not updated since it currently is panicking
      when built with GCC 4.8. Hopefully a new MIPS image will
      appear in a future patch.

Tested by booting Kitkat/ARM and Kitkat/x86 system images in
the emulator and performing basic manual testing.

BUG=16370204

Change-Id: Ia76cf609d181fb3035621d4e34393daaad58d247