Upgrade qemu kernels to support multinetwork.
This patch updates the binaries of emulator-specific kernels to
the following revisions (https://android.googlesource.com/kernel/goldfish):

  android-goldfish-3.4:   ea97df6 goldfish: Enable multinetwork kernel features.
  android-goldfish-3.10:  64a0393 goldfish: Enable multinetwork kernel feature.

The main goals is to get multinetwork working, which is required to
properly emulate Android L and beyond. This also includes other patches,
including compatibility-related patches for the ARM64 kernel that make
ARMv7 binaries run properly under it.

BUG=17516983

Change-Id: I2f1c352b72ca6b254dbcb9732dee3c014867ae99