Fix Mips64 build and --cross option handling.
The Mips64 cross-toolchain is now located under:

  prebuilts/gcc/$SYSTEM/mips/mips64el-linux-android-4.9

So fix the script to pick it from here.

Also, when using --cross=<binprefix>, the option would not
work due to a missing assignment (this is useful when trying
to rebuilt the 2.6.29 kernel with this script, which requires
checking out an old 4.6 cross-toolchain and using --cross to
use it).

Change-Id: Ia98808ac3e95a77c750380d55e76158fbdda4a95